requirements.txt
streamlit
spacy
pdfplumber
python-docx
groq
pandas
python-dotenv
fpdf
